+++ src/app/pages/list/list-layout-popup/import-input-box/import-input-box.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-import-input-box',
+  templateUrl: './import-input-box.component.html',
+  styleUrls: ['./import-input-box.component.scss']
+})
+export class ImportInputBoxComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}