+++ src/app/pages/list/list.module.ts
@@ -32,6 +32,11 @@ import {ListInventoryComponent} from './list-inventory/list-inventory.component'
 import {FlexLayoutModule} from '@angular/flex-layout';
 import {ListTagsPopupComponent} from './list-tags-popup/list-tags-popup.component';
 import {MaintenanceGuard} from '../maintenance/maintenance.guard';
+import { ListHelpComponent } from './list-help/list-help.component';
+import { ListLayoutPopupComponent } from './list-layout-popup/list-layout-popup.component';
+import { ListLayoutRowComponent } from './list-layout-popup/list-layout-row/list-layout-row.component';
+import { ImportInputBoxComponent } from './list-layout-popup/import-input-box/import-input-box.component';
+import {ClipboardModule} from 'ngx-clipboard/dist';
 
 const routes: Routes = [
     {
@@ -73,6 +78,7 @@ const routes: Routes = [
         MatTooltipModule,
 
         FlexLayoutModule,
+        ClipboardModule,
 
         CoreModule,
         PricingModule,
@@ -89,11 +95,18 @@ const routes: Routes = [
         ListNoteComponent,
         ListInventoryComponent,
         ListTagsPopupComponent,
+        ListHelpComponent,
+        ListLayoutPopupComponent,
+        ListLayoutRowComponent,
+        ImportInputBoxComponent,
     ],
     entryComponents: [
         RegenerationPopupComponent,
         TimerOptionsPopupComponent,
-        ListTagsPopupComponent
+        ListTagsPopupComponent,
+        ListHelpComponent,
+        ListLayoutPopupComponent,
+        ImportInputBoxComponent,
     ]
 })
 export class ListModule {