+++ src/app/pages/profile/profile-help/profile-help.component.ts
@@ -1,0 +1,13 @@
+import {Component} from '@angular/core';
+
+@Component({
+    selector: 'app-profile-help',
+    templateUrl: './profile-help.component.html',
+    styleUrls: ['./profile-help.component.scss']
+})
+export class ProfileHelpComponent {
+
+    constructor() {
+    }
+
+}