+++ src/app/pages/recipes/recipes-help/recipes-help.component.ts
@@ -1,0 +1,14 @@
+import {Component} from '@angular/core';
+import {MatDialogRef} from '@angular/material';
+
+@Component({
+    selector: 'app-recipes-help',
+    templateUrl: './recipes-help.component.html',
+    styleUrls: ['./recipes-help.component.scss']
+})
+export class RecipesHelpComponent {
+
+    constructor() {
+    }
+
+}